Set void(0) onclick handlers for fix hover on iOS
On iOS Safarai (and possibly other iOS browsers), the :hover
selector won't work by default because the touch gets turned
into a click event instead. Setting a void(0) onclick handler
ensures we get a hover event. This needs to be set both on
the <div> for the nav menu (to activate the menu popup),
and on the other main <div>s in the page (to de-activate
the menu popup by touching elsewhere).

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>