Add more content to make the existence of labs in step 2 more apparent
Reorganized the introduction to show that the user will go through
a two step process. Step 1 is launching an EC2 instance through the
AWS console. Step 2 is a walkthrough of two labs to help them
get more familiar with KubeVirt.

Also explicitly say they have to navigate back to this page after
launching the console to continue with Step 2.