Add deprecation warning
Since the KubeVirt documentation pages that are part of the KubeVirt Web
site do not fulfill technical requirements given by the KubeVirt
developers, they need to be replaced by an alternative solution.  After
a long discussion we have decided for the AsciiDoc-based solution that
is currently in the phase of implementation.  In the meantime, everybody
should use the 'KubeVirt User Guide' [1] as the only up-to-date
documentation resource.  We should add a at least a deprecation warning
to the first page for convenience.

[1] http://kubevirt.io/user-guide
[1] http://docs.kubevirt.io