Template labs commands into variables
Various parts of lab commands are templated so that we can use the
variables that feed into the templates as a source. The variables
are stored in a yaml file.

We can then use this yaml file in cloud-image-builder to create tests
for the labs. The cloud-image-builder tests can then test against
the same manifests and commands that the users go through in
completing the labs.