Update the copyright line in the license file
The LICENSE file in the website was inherited verbatim from the
generator tool that is used to create the site (jekyll). However, the
content in the site comes from KubeVirt contributors. As the copyright
note is about the web site content and not the tool used to generate it,
it should reflect this.

Signed-off-by: Pep Turró Mauri <pep@redhat.com>