Add custom user agent to HTMLProofer (#660)
* Add custom user agent to HTMLProofer

Link checking on Katacoda is failing with 403 Forbidden. Testing with curl
suggests that a custom user agent prevents this.

Signed-off-by: Pep Turró Mauri <pep@redhat.com>

* Workaround katacoda links failing the link checker

Some links to katacoda.com are causing the link checker to fail with 403
Forbidden messages that don't reflect reality. This is blocking PRs and site
builds, so adding a workaround here by ignoring this domain.

Signed-off-by: Pep Turró Mauri <pep@redhat.com>